+++ resources/views/admin/comms/announcements/index.blade.php
@@ -48,3 +48,7 @@
     @include('admin.comms.announcements._form', ['mode' => 'edit', 'a' => $a])
   @endforeach
 @endsection
+
+@push('scripts')
+  @include('admin.partials.translation-suggest-script')
+@endpush